requirements.txt
websockets>=12.0
aiohttp>=3.9
fastapi>=0.110
uvicorn>=0.27
jinja2>=3.1
scipy>=1.12
numpy>=1.26
pydantic>=2.6
pydantic-settings>=2.1
python-dotenv>=1.0
aiosqlite>=0.20
rich>=13.7
py-clob-client>=0.6
